Updated U128 / I128 and SMART types
The U128 and I128 types had a bug, when passed to a functions as copy (not a reference), the original variable resources would be released when the calling function returned. This has been fixed. This was the same kind of bug the SMART string type had. Moved the definition of the U128 and I128 types above the SMART type definition which allowed automatic reading of those types. Added an example for the U128/I128 types to the ExampleApp.